pyproject.toml
[tool.poetry]
name = "neo4j-blue-green-arrow-etl"
version = "0.1.0"
description = "Blue/Green deployment demo using Neo4j Arrow loader with database aliases"
authors = ["Alex Woolford alex@woolford.io"]
readme = "README.md"
packages = [{include = "blue_green_etl", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pyarrow = "^15.0.0"
neo4j = "^5.20.0"
graphdatascience = "^1.8.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pre-commit = "^3.6.0"
detect-secrets = "^1.4.0"
# black = "^23.0.0"
# flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# Optional: Add any command-line scripts here
